backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
sqlalchemy
jinja2
python-multipart